UI和CNN分离。
